getting closer to being able to run the test suite...

+++ generic/tclDecls.h
@@ -270,10 +270,10 @@ EXTERN int		Tcl_CommandComplete(const char *cmd);
 /* 83 */
 EXTERN char *		Tcl_Concat(size_t argc, const char *const *argv);
 /* 84 */
-EXTERN int		Tcl_ConvertElement(const char *src, char *dst,
+EXTERN size_t		Tcl_ConvertElement(const char *src, char *dst,
 				int flags);
 /* 85 */
-EXTERN int		Tcl_ConvertCountedElement(const char *src,
+EXTERN size_t		Tcl_ConvertCountedElement(const char *src,
 				size_t length, char *dst, int flags);
 /* 86 */
 EXTERN int		Tcl_CreateAlias(Tcl_Interp *slave,
@@ -652,9 +652,9 @@ EXTERN void		Tcl_Release(ClientData clientData);
 /* 217 */
 EXTERN void		Tcl_ResetResult(Tcl_Interp *interp);
 /* 218 */
-EXTERN int		Tcl_ScanElement(const char *src, int *flagPtr);
+EXTERN size_t		Tcl_ScanElement(const char *src, int *flagPtr);
 /* 219 */
-EXTERN int		Tcl_ScanCountedElement(const char *src,
+EXTERN size_t		Tcl_ScanCountedElement(const char *src,
 				size_t length, int *flagPtr);
 /* 220 */
 EXTERN int		Tcl_SeekOld(Tcl_Channel chan, int offset, int mode);
@@ -1923,8 +1923,8 @@ typedef struct TclStubs {
     int (*tcl_Close) (Tcl_Interp *interp, Tcl_Channel chan); /* 81 */
     int (*tcl_CommandComplete) (const char *cmd); /* 82 */
     char * (*tcl_Concat) (size_t argc, const char *const *argv); /* 83 */
-    int (*tcl_ConvertElement) (const char *src, char *dst, int flags); /* 84 */
-    int (*tcl_ConvertCountedElement) (const char *src, size_t length, char *dst, int flags); /* 85 */
+    size_t (*tcl_ConvertElement) (const char *src, char *dst, int flags); /* 84 */
+    size_t (*tcl_ConvertCountedElement) (const char *src, size_t length, char *dst, int flags); /* 85 */
     int (*tcl_CreateAlias) (Tcl_Interp *slave, const char *slaveCmd, Tcl_Interp *target, const char *targetCmd, size_t argc, const char *const *argv); /* 86 */
     int (*tcl_CreateAliasObj) (Tcl_Interp *slave, const char *slaveCmd, Tcl_Interp *target, const char *targetCmd, size_t objc, Tcl_Obj *const objv[]); /* 87 */
     Tcl_Channel (*tcl_CreateChannel) (const Tcl_ChannelType *typePtr, const char *chanName, ClientData instanceData, int mask); /* 88 */
@@ -2065,8 +2065,8 @@ typedef struct TclStubs {
     void (*tcl_RegExpRange) (Tcl_RegExp regexp, int index, const char **startPtr, const char **endPtr); /* 215 */
     void (*tcl_Release) (ClientData clientData); /* 216 */
     void (*tcl_ResetResult) (Tcl_Interp *interp); /* 217 */
-    int (*tcl_ScanElement) (const char *src, int *flagPtr); /* 218 */
-    int (*tcl_ScanCountedElement) (const char *src, size_t length, int *flagPtr); /* 219 */
+    size_t (*tcl_ScanElement) (const char *src, int *flagPtr); /* 218 */
+    size_t (*tcl_ScanCountedElement) (const char *src, size_t length, int *flagPtr); /* 219 */
     int (*tcl_SeekOld) (Tcl_Channel chan, int offset, int mode); /* 220 */
     int (*tcl_ServiceAll) (void); /* 221 */
     int (*tcl_ServiceEvent) (int flags); /* 222 */
